feat/#91 추천 알고리즘 API

+++ src/main/java/org/sopt/certi_server/domain/admin/dto/response/CertificationJobSimple.java
@@ -2,7 +2,7 @@ package org.sopt.certi_server.domain.admin.dto.response;
 
 import org.sopt.certi_server.domain.certification.entity.CertificationJob;
 
-public record CertificationJobSimple(Long certificationJobId, Long jobId, String jobName, float weight) {
+public record CertificationJobSimple(Long certificationJobId, Long jobId, String jobName, double weight) {
 
     public static CertificationJobSimple from(CertificationJob cj){
         return new CertificationJobSimple(cj.getId(), cj.getJob().getId(), cj.getJob().getName(), cj.getWeight());
